Adjust which-key config to match commands
Instead of key sequences, because the user might change the bindings.

Bind paging key for which-key

Also disables special-key handling which is disorienting for those who are not aware of it

Remove out advice on which-key--update: Was not allowing popup with prefix arguments, like SPC u 1 SPC.

Make which-key descriptions respect the users leader key.