Fix inputting backtick and single quote in markdown mode
Not sure why that part of code was introduced in the first place but it seems to no longer work with the newest packages. Both backtick and quote behaviors were a kind of broken. Removing the code makes them normal again. And there seems to be nothing wrong with Github code blocks.