Make Y equivalent to y$.
This is much more consistent with C and D and is even recommended by vim
itself.

The ~Y~ key does not yank the whole line. It yanks from the current
point to the end of the line. This is more consistent with the behavior
of ~C~ and ~D~ and is also recommended by the vim documentation.