Improve keybindings for text-region counting.
Emacs provides a `count-words-region' which is badly named as it counts not only words, but also characters and lines. The current keybinding for this function `SPC x w c' implies it's referring to words only. The new mapping `SPC x c` is more coherent since it means counting text, whether it is characters, words or lines. An alias `count-region` is created to be less confusing in which-key description.

`SPC x w c` being free, `spacemacs/count-words-analysis` is remapped to it instead of `SPC x w C` to save one keystroke.