remove rainbow-delimiters-mode from specific hooks
handle `rainbow-delimiters-mode` by adding it to the `prog-mode-hook`,
if wanted by the user. some layers are adding it on their own mode-hook,
having for effect that `rainbow-delimiters-mode` is called twice, which
disable it.

this commit remove these layer-specific definitions of
`rainbow-delimiters` as it is now handled by the `prog-mode-defaults`.
It also takes care of running `prog-mode-hook` in modes that are not
derived from it.